project01/requirements.txt
alabaster==0.7.10
anaconda-client==1.6.5
anaconda-navigator==1.6.9
anaconda-project==0.8.0
APScheduler==3.6.0
asn1crypto==0.22.0
astroid==1.5.3
astropy==2.0.2
babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.0.0
blis==0.2.4
bokeh==0.12.10
boto==2.48.0
boto3==1.9.205
botocore==1.12.205
Bottleneck==1.2.1
CacheControl==0.12.3
certifi==2019.6.16
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.4.0
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.2
conda==4.3.30
conda-build==3.0.27
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.0.3
cycler==0.10.0
cymem==2.0.2
Cython==0.26.1
cytoolz==0.8.2
dask==0.15.3
datashape==0.5.4
decorator==4.1.2
distlib==0.2.5
distributed==1.19.1
Django==2.2.4
docopt==0.6.2
docutils==0.14
en-core-web-sm==2.1.0
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.12
Flask==0.12.2
Flask-Cors==3.0.3
gensim==3.8.0
gevent==1.2.2
glob2==0.5
greenlet==0.4.12
h5py==2.7.0
happybase==1.1.0
heapdict==1.0.0
html5lib==0.999999999
idna==2.8
imageio==2.2.0
imagesize==0.7.1
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
isort==4.2.15
itchat==1.3.10
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
jieba==0.39
Jinja2==2.9.6
jmespath==0.9.4
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyterlab==0.27.0
jupyterlab-launcher==0.4.0
Lasagne==0.1
lazy-object-proxy==1.3.1
llvmlite==0.20.0
locket==0.2.0
lockfile==0.12.2
lxml==4.1.0
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
menuinst==1.4.10
mistune==0.7.4
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
murmurhash==1.0.2
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
nltk==3.2.4
nose==1.3.7
notebook==5.0.0
numba==0.35.0+10.g143f70e
numexpr==2.6.2
numkit==1.1.0
numpy==1.16.3
numpydoc==0.7.0
odo==0.5.1
olefile==0.44
opencc-python-reimplemented==0.1.4
openpyxl==2.4.8
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
partd==0.3.8
path.py==10.3.1
pathlib2==2.3.0
patsy==0.4.1
pep8==1.7.0
pickleshare==0.7.4
Pillow==4.2.1
pipreqs==0.4.9
pkginfo==1.4.1
plac==0.9.6
ply==3.10
preshed==2.0.1
progress==1.3
prompt-toolkit==1.0.15
psutil==5.4.0
py==1.4.34
pycodestyle==2.3.1
pycosat==0.6.2
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.7.4
pyltp==0.2.1
PyMySQL==0.9.3
pyodbc==4.0.17
pyOpenSSL==17.2.0
pyparsing==2.2.0
pypng==0.0.19
PyQRCode==1.2.1
PySocks==1.6.7
pytest==3.2.1
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
pywin32==221
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.22.0
rope==0.10.5
ruamel-yaml==0.11.14
s3transfer==0.2.1
scikit-image==0.13.0
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.8.4
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
spacy==2.1.4
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.4
SQLAlchemy==1.1.13
sqlparse==0.3.0
srsly==0.0.5
stanfordcorenlp==3.9.1.1
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
testpath==0.3.1
Theano==1.0.4
thinc==7.0.4
thriftpy==0.3.9
toolz==0.8.2
tornado==4.5.2
tqdm==4.32.1
traitlets==4.3.2
typing==3.6.2
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.3
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.2
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.4
xlwt==1.3.0
xx-ent-wiki-sm==2.1.0
yarg==0.1.9
zhon==1.1.5
zict==0.1.3
